pyproject.toml
[project]
name = "dq-runner"
version = "0.1.0"
description = "Config-driven data quality runner built on Great Expectations and Pandas."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "great-expectations>=1.0.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
dq-runner = "dq_runner.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["dq_runner"]
